+++ ownCloud/Server List/ServerListTableViewController.swift
@@ -728,7 +728,7 @@ class ServerListTableViewController: UITableViewController, Themeable, StateRest
 			}
 
 			if tableView.isEditing {
-				self.showBookmarkUI(edit: bookmark)
+				self.showBookmarkUI(edit: bookmark, removeAuthDataFromCopy: false)
 			} else {
 				self.connect(to: bookmark, lastVisibleItemId: nil, animated: true)
 				self.tableView.deselectRow(at: indexPath, animated: true)
@@ -762,7 +762,7 @@ class ServerListTableViewController: UITableViewController, Themeable, StateRest
 			menuItems.append(openWindow)
 		}
 		let edit = UIAction(title: "Edit".localized, image: UIImage(systemName: "gear")) { _ in
-			self.showBookmarkUI(edit: bookmark)
+			self.showBookmarkUI(edit: bookmark, removeAuthDataFromCopy: false)
 		}
 		if VendorServices.shared.canEditAccount {
 			menuItems.append(edit)
@@ -860,7 +860,7 @@ class ServerListTableViewController: UITableViewController, Themeable, StateRest
 
 		let editRowAction = UIContextualAction(style: .normal, title: "Edit".localized, handler: { [weak self] (_, _, completionHandler) in
 			if let bookmark = OCBookmarkManager.shared.bookmark(at: UInt(indexPath.row)) {
-				self?.showBookmarkUI(edit: bookmark)
+				self?.showBookmarkUI(edit: bookmark, removeAuthDataFromCopy: false)
 			}
 			completionHandler(true)
 		})