+++ ownCloudAppFramework/Building/BuildOptions.m
@@ -23,6 +23,8 @@ OCClassSettingsIdentifier OCClassSettingsIdentifierBuildOptions = @"build";
 OCClassSettingsKey	  OCClassSettingsKeyBuildFlags = @"flags";
 OCClassSettingsKey	  OCClassSettingsKeyCustomAppScheme = @"custom-app-scheme";
 OCClassSettingsKey	  OCClassSettingsKeyCustomAuthScheme = @"custom-auth-scheme";
+OCClassSettingsKey	  OCClassSettingsKeyAppGroupIdentifier = @"app-group-identifier";
+OCClassSettingsKey	  OCClassSettingsKeyOCAppGroupIdentifier = @"oc-app-group-identifier";
 
 @implementation BuildOptions
 
@@ -51,7 +53,7 @@ OCClassSettingsKey	  OCClassSettingsKeyCustomAuthScheme = @"custom-auth-scheme";
 		// build.flags
 		OCClassSettingsKeyBuildFlags : @{
 			OCClassSettingsMetadataKeyType 		: OCClassSettingsMetadataTypeString,
-			OCClassSettingsMetadataKeyDescription 	: @"A set of space separated flags to customize the build. Must be provided in Branding.plist at build time. For documentation, please see doc/BUILD_CUSTOMIZATION.md.",
+			OCClassSettingsMetadataKeyDescription 	: @"A set of space separated flags to customize the build. Must be provided in Branding.plist at build time. For documentation, please see https://github.com/owncloud/ios-app/blob/master/doc/BUILD_CUSTOMIZATION.md.",
 			OCClassSettingsMetadataKeyStatus	: OCClassSettingsKeyStatusSupported,
 			OCClassSettingsMetadataKeyCategory	: @"Build",
 		},
@@ -59,7 +61,7 @@ OCClassSettingsKey	  OCClassSettingsKeyCustomAuthScheme = @"custom-auth-scheme";
 		// build.custom-app-scheme
 		OCClassSettingsKeyCustomAppScheme : @{
 			OCClassSettingsMetadataKeyType 		: OCClassSettingsMetadataTypeString,
-			OCClassSettingsMetadataKeyDescription 	: @"Name of the URL scheme to use for private links. Must be provided in Branding.plist at build time. For documentation, please see doc/BUILD_CUSTOMIZATION.md.",
+			OCClassSettingsMetadataKeyDescription 	: @"Name of the URL scheme to use for private links. Must be provided in Branding.plist at build time. For documentation, please see https://github.com/owncloud/ios-app/blob/master/doc/BUILD_CUSTOMIZATION.md.",
 			OCClassSettingsMetadataKeyStatus	: OCClassSettingsKeyStatusSupported,
 			OCClassSettingsMetadataKeyCategory	: @"Build",
 		},
@@ -67,7 +69,23 @@ OCClassSettingsKey	  OCClassSettingsKeyCustomAuthScheme = @"custom-auth-scheme";
 		// build.custom-app-scheme
 		OCClassSettingsKeyCustomAuthScheme : @{
 			OCClassSettingsMetadataKeyType 		: OCClassSettingsMetadataTypeString,
-			OCClassSettingsMetadataKeyDescription 	: @"Name of the URL scheme to use for OAuth2/OIDC authentication. Must be provided in Branding.plist at build time. The authentication redirect URI parameters must also be changed accordingly in Branding.plist and on the server side. For documentation, please see doc/BUILD_CUSTOMIZATION.md.",
+			OCClassSettingsMetadataKeyDescription 	: @"Name of the URL scheme to use for OAuth2/OIDC authentication. Must be provided in Branding.plist at build time. The authentication redirect URI parameters must also be changed accordingly in Branding.plist and on the server side. For documentation, please see https://github.com/owncloud/ios-app/blob/master/doc/BUILD_CUSTOMIZATION.md.",
+			OCClassSettingsMetadataKeyStatus	: OCClassSettingsKeyStatusSupported,
+			OCClassSettingsMetadataKeyCategory	: @"Build",
+		},
+
+		// build.app-group-identifier
+		OCClassSettingsKeyAppGroupIdentifier : @{
+			OCClassSettingsMetadataKeyType 		: OCClassSettingsMetadataTypeString,
+			OCClassSettingsMetadataKeyDescription 	: @"Set a custom app group identifier via Branding.plist parameter. This value will be set by fastlane. Changes OCAppGroupIdentifier, OCKeychainAccessGroupIdentifier and updates other, directly signing-relevant parts of the Info.plist. With this value set, fastlane needs the provisioning profiles and certificate with the app group identifier. This is needed, if a customer is using an own resigning script which does not handle setting the app group identifier.",
+			OCClassSettingsMetadataKeyStatus	: OCClassSettingsKeyStatusSupported,
+			OCClassSettingsMetadataKeyCategory	: @"Build",
+		},
+
+		// build.oc-app-group-identifier
+		OCClassSettingsKeyOCAppGroupIdentifier : @{
+			OCClassSettingsMetadataKeyType 		: OCClassSettingsMetadataTypeString,
+			OCClassSettingsMetadataKeyDescription 	: @"Set a custom app group identifier via Branding.plist parameter. This value will be set by fastlane. Changes OCAppGroupIdentifier, OCKeychainAccessGroupIdentifier only. Fastlane does not need the provisioning profile and certificate with the given app group identifer. Needs resigning with the correct provisioning profile and certificate. This is needed, if a customer is using an own resigning script which does not handle setting the app group identifier.",
 			OCClassSettingsMetadataKeyStatus	: OCClassSettingsKeyStatusSupported,
 			OCClassSettingsMetadataKeyCategory	: @"Build",
 		},