+++ ownCloudAppShared/User Interface/Theme/UI/ThemeTableViewCell.swift
@@ -191,6 +191,10 @@ open class ThemeTableViewCell: UITableViewCell, Themeable {
 						textColor = collection.tableRowColors.labelColor
 						backgroundColor = collection.tableRowColors.backgroundColor
 
+					case .text:
+						textColor = collection.tableRowColors.labelColor
+						backgroundColor = collection.tableRowColors.backgroundColor
+
 					case .confirmation:
 						textColor = collection.approvalColors.normal.foreground
 						backgroundColor = collection.approvalColors.normal.background