+++ CHANGELOG.md
@@ -1,3 +1,140 @@
+Changelog for ownCloud iOS Client [11.9.1] (2022-03-29)
+=======================================
+The following sections list the changes in ownCloud iOS Client 11.9.1 relevant to
+ownCloud admins and users.
+
+[11.9.1]: https://github.com/owncloud/ios-app/compare/milestone/11.9.0...milestone/11.9.1
+
+Summary
+-------
+
+* Bugfix - Setup Passcode with Biometrical Unlock: [#1112](https://github.com/owncloud/ios-app/pull/1112)
+* Change - Set App Group Identifier: [#1099](https://github.com/owncloud/ios-app/pull/1099)
+
+Details
+-------
+
+* Bugfix - Setup Passcode with Biometrical Unlock: [#1112](https://github.com/owncloud/ios-app/pull/1112)
+
+   Biometrical unlock button no longer appear in setup view and after passcode was successfully
+   setup, show biometrical unlock for permissions dialog.
+
+   https://github.com/owncloud/ios-app/pull/1112
+
+* Change - Set App Group Identifier: [#1099](https://github.com/owncloud/ios-app/pull/1099)
+
+   Set a custom app group identifier via Branding.plist this parameter. This value will be set by
+   fastlane to all needed Info.plist keys. This is needed, if a customer is using an own resigning
+   script which does not handle setting the app group identifier.
+
+   https://github.com/owncloud/ios-app/pull/1099
+
+Changelog for ownCloud iOS Client [11.9.0] (2022-03-16)
+=======================================
+The following sections list the changes in ownCloud iOS Client 11.9.0 relevant to
+ownCloud admins and users.
+
+[11.9.0]: https://github.com/owncloud/ios-app/compare/milestone/11.8.2...milestone/11.9.0
+
+Summary
+-------
+
+* Bugfix - Fix WebDAV endpoint URL for media playback after restoration: [#1093](https://github.com/owncloud/ios-app/pull/1093)
+* Bugfix - OAuth token renewal race condition: [#1105](https://github.com/owncloud/ios-app/pull/1105)
+* Change - Biometrical Authentication Button: [#1004](https://github.com/owncloud/ios-app/issues/1004)
+* Change - Poll for changes efficiency enhancements: [#1043](https://github.com/owncloud/ios-app/pull/1043)
+* Change - Webfinger / server location: [#1059](https://github.com/owncloud/ios-app/pull/1059)
+* Change - Infinite PROPFIND support: [#950](https://github.com/owncloud/ios-app/issues/950)
+* Change - Rename Account (without re-authentication): [#972](https://github.com/owncloud/ios-app/issues/972)
+
+Details
+-------
+
+* Bugfix - Fix WebDAV endpoint URL for media playback after restoration: [#1093](https://github.com/owncloud/ios-app/pull/1093)
+
+   Fixes a bug where media playback failed with a 404 Not Found error after restoration because the
+   WebDAV endpoint URL was constructed from authentication data rather than OC user endpoint
+   data.
+
+   https://github.com/owncloud/ios-app/pull/1093
+
+* Bugfix - OAuth token renewal race condition: [#1105](https://github.com/owncloud/ios-app/pull/1105)
+
+   Retry requests that failed with a 401 during a token refresh
+
+   https://github.com/owncloud/ios-app/pull/1105
+
+* Change - Biometrical Authentication Button: [#1004](https://github.com/owncloud/ios-app/issues/1004)
+
+   Added biometrical authentication button to provide a fallback for the fileprovider or app, if
+   the automatically biometrical unlock does not work, or the user cancel the biometrical
+   authentication flow.
+
+   https://github.com/owncloud/ios-app/issues/1004
+
+* Change - Poll for changes efficiency enhancements: [#1043](https://github.com/owncloud/ios-app/pull/1043)
+
+   Avoids simultaneous polling for changes by FileProvider and app.
+
+   https://github.com/owncloud/ios-app/pull/1043
+
+* Change - Webfinger / server location: [#1059](https://github.com/owncloud/ios-app/pull/1059)
+
+   Allows using webfinger or a lookup table to locate and use an alternative server based on the
+   user name
+
+   https://github.com/owncloud/ios-app/pull/1059
+
+* Change - Infinite PROPFIND support: [#950](https://github.com/owncloud/ios-app/issues/950)
+
+   Added support for prepopulation of newly created account bookmarks via infinite PROPFINDs,
+   which speeds up the initial scan
+
+   https://github.com/owncloud/ios-app/issues/950
+
+* Change - Rename Account (without re-authentication): [#972](https://github.com/owncloud/ios-app/issues/972)
+
+   Check if only the account name was changed in edit mode: save and dismiss without
+   re-authentication
+
+   https://github.com/owncloud/ios-app/issues/972
+
+Changelog for ownCloud iOS Client [11.8.2] (2022-01-17)
+=======================================
+The following sections list the changes in ownCloud iOS Client 11.8.2 relevant to
+ownCloud admins and users.
+
+[11.8.2]: https://github.com/owncloud/ios-app/compare/milestone/11.8.1...milestone/11.8.2
+
+Summary
+-------
+
+* Bugfix - Continuous Audio Playback: [#4924](https://github.com/owncloud/enterprise/issues/4924)
+* Bugfix - PDF Editing: [#4934](https://github.com/owncloud/enterprise/issues/4934)
+* Change - (Branding) Corporate Color as Folder Color: [#1069](https://github.com/owncloud/ios-app/issues/1069)
+
+Details
+-------
+
+* Bugfix - Continuous Audio Playback: [#4924](https://github.com/owncloud/enterprise/issues/4924)
+
+   Fixed continuous audio playback, which stopped after two audio files.
+
+   https://github.com/owncloud/enterprise/issues/4924
+
+* Bugfix - PDF Editing: [#4934](https://github.com/owncloud/enterprise/issues/4934)
+
+   Fixed bug that prevents changes to PDFs being saved in place.
+
+   https://github.com/owncloud/enterprise/issues/4934
+
+* Change - (Branding) Corporate Color as Folder Color: [#1069](https://github.com/owncloud/ios-app/issues/1069)
+
+   Use the corporate color as folder color as default color (can be overridden by the specific
+   key/value pair).
+
+   https://github.com/owncloud/ios-app/issues/1069
+
 Changelog for ownCloud iOS Client [11.8.1] (2021-12-22)
 =======================================
 The following sections list the changes in ownCloud iOS Client 11.8.1 relevant to